refactor(cost-analysis-data-table): modify helper location
